refactor(Client): 优化 NvidiaGpuController 中的参数传递方式，改为逐项传递以避免参数注入风险；调整 AutoStartController 以支持笔记本电池场景
