+++ JiaoLongControl/Server/Core/Utils/SelfStart.cs
@@ -1,4 +1,3 @@
-using JiaoLongControl.Server.Core.Models;
 using JiaoLongControl.Server.Interop;
 
 namespace JiaoLongControl.Server.Core.Utils;
